Restructured view.
Since navigating with a specified calendar is now obsolete, the controller only uses a user to navitate.

If the user is the connected user, you'll get the usual calendar view with all his calendar and the calendars of a contact.
If the user is an other one, you'll get a union of all his calendars.